+++ lib/components/index.d.ts
@@ -10,6 +10,7 @@ export * from './Drawer';
 export * from './DropDown';
 export * from './ExpansionPanel';
 export * from './FileInput';
+export * from './FloatingActionButton';
 export * from './Form';
 export * from './Grid';
 export * from './IconButton';